Backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
asn1crypto==1.5.1
attrs==25.3.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
dataclasses-json==0.5.9
distro==1.9.0
fastapi==0.115.0
filelock==3.18.0
filetype==1.2.0
fonttools==4.59.0
frozenlist==1.7.0
fsspec==2025.7.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.178.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.4
idna==3.10
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.9
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.74
langchain-experimental==0.3.4
langchain-google-genai==2.0.10
langchain-openai==0.3.29
langchain-text-splitters==0.3.9
langsmith==0.4.13
marshmallow==3.26.1
marshmallow-enum==1.5.1
matplotlib==3.9.2
multidict==6.6.4
mypy_extensions==1.1.0
numexpr==2.11.0
numpy==2.3.2
openai==1.99.9
openapi-schema-pydantic==1.2.4
orjson==3.11.1
packaging==25.0
pandas==2.2.3
pg8000==1.31.2
pillow==11.3.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
psycopg==3.2.9
psycopg-binary==3.2.9
psycopg-pool==3.2.6
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.4
requests-toolbelt==1.0.0
rsa==4.9.1
safetensors==0.6.2
scipy==1.16.1
scramp==1.4.6
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.38.6
tenacity==9.1.2
tiktoken==0.11.0
tokenizers==0.21.4
tqdm==4.67.1
transformers==4.55.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.30.6
yarl==1.20.1
zstandard==0.23.0
